requirements.txt
asgi-redis==1.0.0
asgiref==1.0.0
autobahn==0.17.1
channels==1.0.2
constantly==15.1.0
daphne==1.0.1
Django==1.10.5
django-braces==1.10.0
django-cors-headers==2.0.1
django-oauth-toolkit==0.11.0
djangorestframework==3.5.3
djangorestframework-oauth==1.1.0
incremental==16.10.1
msgpack-python==0.4.8
oauthlib==1.1.2
posix-ipc==1.0.0
redis==2.10.5
six==1.10.0
Twisted==16.6.0
txaio==2.6.0
zope.interface==4.3.3
